chore: split wrapper and settingsWrapper
